第二版本

+++ src/main/java/com/wiki/core/common/reponse/PageDataResponse.java
@@ -10,17 +10,17 @@ import lombok.Data;
 @Data
 public class PageDataResponse<T> extends DataResponse<T>{
 
-    private Integer page;
+    private Long page;
 
-    private Integer size;
+    private Long size;
 
-    private Integer total;
+    private Long total;
 
     public PageDataResponse(T t) {
         super(t);
     }
 
-    public static <T> PageDataResponse<T> of(T data,Integer page,Integer size,Integer total){
+    public static <T> PageDataResponse<T> of(T data,Long page,Long size,Long total){
         PageDataResponse<T> response = new PageDataResponse<>(data);
         response.setPage(page);
         response.setSize(size);
